refactor: change navbar component to pages component. Design project changed and navbar hadn't sense on this updated context
