+++ trilha-thomas/src/app/app.component.html
@@ -1,4 +1,4 @@
 <div class="body">
   <app-header></app-header>
-  <app-navbar></app-navbar>
+  <pages></pages>
 </div>
