--- trilha-thomas/src/app/components/navbar/navbar.component.html
@@ -1,16 +1,0 @@
-<section class="container nav-section">
-  <nav class="navbar">
-    <ul>
-      <li>
-        <a routerLink="/motivation">motivation</a> <br />
-        <a routerLink="/about">about</a> <br />
-        <a routerLink="/intersts">intersts</a> <br />
-        <a routerLink="/coders-users"> coders-users</a> <br />
-      </li>
-    </ul>
-  </nav>
-
-  <div class="routers-render">
-    <router-outlet></router-outlet>
-  </div>
-</section>
