--- trilha-thomas/src/app/components/navbar/navbar.component.scss
@@ -1,10 +1,0 @@
-.container {
-  max-width: 80%;
-  margin: 0 auto;
-}
-.nav-section {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-}
