+++ trilha-thomas/src/app/components/pages/about/about.component.ts
