+++ trilha-thomas/src/app/components/pages/home/home.component.ts
