+++ trilha-thomas/src/app/components/pages/pages.component.html
@@ -1,0 +1,5 @@
+<section class="container-fluid nav-section">
+  <div class="routers-render container-fluid">
+    <router-outlet></router-outlet>
+  </div>
+</section>
