+++ trilha-thomas/src/app/components/pages/pages.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'pages',
+  templateUrl: './pages.component.html',
+  styleUrls: ['./pages.component.scss'],
+})
+export class PagesComponent {}
