+++ trilha-thomas/src/app/components/pages/registered-users/registered-users.component.ts
